#790 - Moved logic to pause audio when headphones are unplugged to the AudioService
